+++ src/main/java/br/com/algaworks/algafoods/requersts/RestaurantPostRequestBody.java
@@ -1,7 +1,5 @@
 package br.com.algaworks.algafoods.requersts;
 
-import java.math.BigDecimal;
-
 import br.com.algaworks.algafoods.domain.Kitchen;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class RestaurantPostRequestBody {
 
     private String name;
-    private BigDecimal freight;
+    private Long freight;
 	private Kitchen kitchen;
 
 }
