+++ src/main/java/br/com/algaworks/algafoods/requersts/RestaurantPutRequestBody.java
@@ -1,7 +1,5 @@
 package br.com.algaworks.algafoods.requersts;
 
-import java.math.BigDecimal;
-
 import br.com.algaworks.algafoods.domain.Kitchen;
 import lombok.Builder;
 import lombok.Data;
@@ -12,7 +10,7 @@ public class RestaurantPutRequestBody {
 
 	private Long id;
 	private String name;
-	private BigDecimal freight;
+	private Long freight;
 	private Kitchen kitchen;
 
 }
